pyproject.toml
[project]
name = "oot3dhdtextgenerator"
version = "0.1.0"
description = ""
authors = [
    { name = "Karl Debiec", email = "karl.t.debiec@gmail.com" }
]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.2.6",
    "Pillow>=12.1.0",
    "torch>=2.10.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.25.0",
    "pipescaler @ git+https://github.com/KarlTDebiec/PipeScaler.git"
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.3",
    "ty>=0.0.13",
    "types-Pillow>=10.2.0.20240822",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "N", # pep8-naming
    "NPY", # numpy-specific rules
    "PERF", # perflint
    "PL", # pylint
    "UP", # pyupgrade
    "W", # pycodestyle
]
ignore = [
    "D410", # Missing blank line after section
    "D411", # Missing blank line before section
    "PLR2004" # Magic value used in comparison
]

[tool.ruff.lint.pydocstyle]
convention = 'google'

[tool.setuptools]
packages = { find = { include = ["oot3dhdtextgenerator*"] } }
